refactor: reuse styiling in both viewers
