+++ application/ui/src/features/robots/setup-wizard/shared/setup-robot-viewer.tsx
@@ -2,7 +2,7 @@
 
 import { Suspense, useMemo, useRef } from 'react';
 
-import { Grid, OrbitControls, PerspectiveCamera } from '@react-three/drei';
+import { OrbitControls, PerspectiveCamera } from '@react-three/drei';
 import { Canvas, useFrame, useThree } from '@react-three/fiber';
 import * as THREE from 'three';
 import type { OrbitControls as OrbitControlsImpl } from 'three-stdlib';
@@ -12,6 +12,7 @@ import { URDFRobot } from 'urdf-loader';
 import { useContainerSize } from '../../../../components/zoom/use-container-size';
 import { useLoadModelQuery } from '../../robot-models-context';
 import { SchemaRobotType } from '../../robot-types';
+import { RobotViewerScene, useConfigureModelShadows } from '../../robot-viewer-scene';
 import { JointHighlight, useJointHighlight } from './use-joint-highlight';
 
 // ---------------------------------------------------------------------------
@@ -23,6 +24,7 @@ const ActualURDFModel = ({ model, highlights }: { model: URDFRobot; highlights: 
     const scale = [3, 3, 3] as const;
 
     useJointHighlight(model, highlights);
+    useConfigureModelShadows(model);
 
     return (
         <group rotation={rotation} scale={scale}>
@@ -149,20 +151,11 @@ export const SetupRobotViewer = ({ robotType, highlights = [] }: SetupRobotViewe
     return (
         <div ref={ref} style={{ width: '100%', height: '100%' }}>
             <div className='canvas-container' style={{ height: `${size.height}px`, width: `${size.width}px` }}>
-                <Canvas shadows>
-                    <color attach='background' args={['#242528']} />
-                    <ambientLight intensity={0.5} />
-                    <directionalLight
-                        position={[10, 10, 5]}
-                        intensity={1}
-                        castShadow
-                        shadow-mapSize-width={1024}
-                        shadow-mapSize-height={1024}
-                    />
+                <Canvas>
+                    <RobotViewerScene />
                     <PerspectiveCamera makeDefault position={[2.0, 1, 1]} />
                     <OrbitControls ref={controlsRef} enableDamping={false} />
                     <CameraController controlsRef={controlsRef} model={model} highlights={highlights} />
-                    <Grid infiniteGrid cellSize={0.25} sectionColor={'rgb(0, 199, 253)'} fadeDistance={10} />
                     {model && (
                         <group key={model.uuid} position={[0, 0, 0]} rotation={[0, angle, 0]}>
                             <Suspense fallback={null}>
